--- src/redux/instructor/my-courses/index.js
@@ -1,7 +1,0 @@
-import { createReduxStore, createSelector } from '../../redux-creators';
-import myCoursesReducer from './my-courses.reducers';
-import * as myCoursesActions from './my-courses.actions';
-
-export const connectWithRedux = createSelector(myCoursesActions);
-
-export const myCoursesStore = createReduxStore(myCoursesReducer);
