--- src/redux/instructor/my-courses/my-courses.action.types.js
@@ -1,2 +1,0 @@
-export const SET_TERMS = 'my-c-set-terms';
-export const SET_OFFERINGS = 'my-c-set-offerings';
