--- src/redux/instructor/my-courses/my-courses.state.js
@@ -1,6 +1,0 @@
-import { ARRAY_INIT } from 'utils';
-
-export const initialState = {
-  terms: [],
-  offerings: ARRAY_INIT,
-};
