--- src/screens/Instructor/MyCourses/controllers/index.js
@@ -1,3 +1,0 @@
-export { myCoursesStore, connectWithRedux } from 'redux/instructor/my-courses';
-
-export { setup } from './setup';
